fixed implicit not null
