+++ src/test/java/se/eris/ImplicitNotNullInstrumenterTest.java
@@ -22,6 +22,7 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
 import se.eris.maven.NopLogWrapper;
+import se.eris.notnull.NotNullConfiguration;
 import se.eris.util.ReflectionUtil;
 
 import javax.tools.JavaCompiler;
@@ -38,8 +39,8 @@ import static org.hamcrest.MatcherAssert.assertThat;
 
 public class ImplicitNotNullInstrumenterTest {
 
-    public static final File SRC_DIR = new File("src/test/data");
-    public static final File TARGET_DIR = new File("src/test/data");
+    private static final File SRC_DIR = new File("src/test/data");
+    private static final File TARGET_DIR = new File("src/test/data");
 
     @Rule
     public ExpectedException exception = ExpectedException.none();
@@ -49,46 +50,46 @@ public class ImplicitNotNullInstrumenterTest {
         final String fileToCompile = getSrcFile(SRC_DIR, "se/eris/test/TestNotNull.java");
         compile(fileToCompile);
 
+        final NotNullConfiguration configuration = new NotNullConfiguration(true, Collections.singleton("org.jetbrains.annotations.Nullable"));
         final NotNullInstrumenter instrumenter = new NotNullInstrumenter(new NopLogWrapper());
-        final int numberOfInstrumentedFiles = instrumenter.addNotNullAnnotations("src/test/data/se/eris/test", Collections.<String>emptySet(), Collections.<URL>emptyList());
+        final int numberOfInstrumentedFiles = instrumenter.addNotNullAnnotations("src/test/data/se/eris/test", configuration, Collections.<URL>emptyList());
 
         assertThat(numberOfInstrumentedFiles, is(1));
     }
 
     @Test
-    public void parameterFT() throws Exception {
+    public void notNullAnnotatedParameter_shouldValidate() throws Exception {
         final Class<?> c = getCompiledClass(TARGET_DIR, "se.eris.test.TestNotNull");
         final Method notNullParameterMethod = c.getMethod("notNullParameter", String.class);
         ReflectionUtil.simulateMethodCall(notNullParameterMethod, "should work");
         exception.expect(IllegalArgumentException.class);
-        exception.expectMessage("Argument 0 for @NotNull parameter of se/eris/test/TestNotNull.notNullParameter must not be null");
+        exception.expectMessage("Argument 0 for implicit 'NotNull' parameter of se/eris/test/TestNotNull.notNullParameter must not be null");
         ReflectionUtil.simulateMethodCall(notNullParameterMethod, new Object[]{null});
     }
 
     @Test
-    public void parameter_implicit() throws Exception {
+    public void implicitParameter_shouldValidate() throws Exception {
         final Class<?> c = getCompiledClass(TARGET_DIR, "se.eris.test.TestNotNull");
         final Method implicitParameterMethod = c.getMethod("implicitParameter", String.class);
         exception.expect(IllegalArgumentException.class);
-        exception.expectMessage("Argument 0 for @NotNull parameter of se/eris/test/TestNotNull.implicitParameter must not be null");
+        exception.expectMessage("Argument 0 for implicit 'NotNull' parameter of se/eris/test/TestNotNull.implicitParameter must not be null");
         ReflectionUtil.simulateMethodCall(implicitParameterMethod, new Object[]{null});
     }
 
     @Test
-    public void parameter_nullable() throws Exception {
+    public void nullableAnnotatedParameter_shouldNotValidate() throws Exception {
         final Class<?> c = getCompiledClass(TARGET_DIR, "se.eris.test.TestNotNull");
         final Method implicitParameterMethod = c.getMethod("nullableParameter", String.class);
         ReflectionUtil.simulateMethodCall(implicitParameterMethod, new Object[]{null});
     }
 
-    // todo
     @Test
-    public void returnFT() throws Exception {
+    public void implicitReturn_shouldValidate() throws Exception {
         final Class<?> c = getCompiledClass(TARGET_DIR, "se.eris.test.TestNotNull");
-        final Method notNullReturnMethod = c.getMethod("notNullReturn", String.class);
+        final Method notNullReturnMethod = c.getMethod("implicitReturn", String.class);
         ReflectionUtil.simulateMethodCall(notNullReturnMethod, "should work");
         exception.expect(IllegalStateException.class);
-        exception.expectMessage("@NotNull method se/eris/test/TestNotNull.notNullReturn must not return null");
+        exception.expectMessage("@NotNull method se/eris/test/TestNotNull.implicitReturn must not return null");
         ReflectionUtil.simulateMethodCall(notNullReturnMethod, new Object[]{null});
     }
